broke out to individual files
improve dialog display
