+++ packages/extension-utils/package.json
@@ -35,7 +35,6 @@
   "devDependencies": {
     "@looker/components-test-utils": "^1.5.26",
     "@testing-library/react": "^12.1.5",
-    "@testing-library/user-event": "^14.4.3",
     "@types/react-router": "^5.1.18",
     "@types/redux": "^3.6.0",
     "webpack-bundle-analyzer": "^4.2.0",
